+++ driver/src/pm/xaie_clock.c
@@ -151,7 +151,7 @@ AieRC XAie_PmSetColumnClk(XAie_DevInst *DevInst, u32 StartCol, u32 NumCols,
 		return XAIE_INVALID_ARGS;
 	}
 
-	u32 PartEndCol = DevInst->StartCol + DevInst->NumCols - 1;
+	u32 PartEndCol = (u32)(DevInst->StartCol + DevInst->NumCols - 1U);
 
 	if((StartCol < DevInst->StartCol) || (StartCol > PartEndCol) ||
 	   ((StartCol + NumCols - 1) > PartEndCol) ) {
